+++ tests/testthat/test_extended_seq.context_functions.R
@@ -57,6 +57,8 @@ test_that("extended seq.context functions for Strelka GRCh37 vcf", {
 })
 
 test_that("extended seq.context functions for Strelka GRCm38 vcf", {
+  skip_if("" == system.file(package = "BSgenome.Mmusculus.UCSC.mm10"))
+  stopifnot(requireNamespace("BSgenome.Mmusculus.UCSC.mm10"))
   vcf <- ReadAndSplitStrelkaSBSVCFs("testdata/Strelka.SBS.GRCm38.vcf")
   sbs.vcf <- vcf$SBS[[1]]
   mat <- CreateOnePPMFromSBSVCF(sbs.vcf, ref.genome = "GRCm38",