+++ NEWS.md
@@ -1,3 +1,7 @@
+# ICAMS 2.0.5
+* Changes in response to initial submission to CRAN.
+* Added support for BSgenome.Mmusculus.UCSC.mm10. 
+
 # ICAMS 2.0
 First stable release. Mutational catalogs are S3 classes.